Use a pool allocator for symbols not in the symbol table.
We don't use the pool allocator for symbols in the symbol table.
We have to maintain an ordered link list of all symbols in the symbol
table, so it's easy enough to reuse that for deleting the symbols.
The memory impact of keeping a second linked list (internal to the pool)
seems to outweigh (in profiling) the moderate time savings of pool deletion.