Reimplement Expr using flat storage (instead of an allocated tree).
Instead of individual Expr nodes being allocated, each with its own terms
vector, the entire Expr is stored as a single vector that includes both
operators and terms with depth information.  There are examples of this in
the Expr class Doxygen documentation.

This reduces the number of allocations by 50% and executed instruction count
by 10-15% on large input files.

It also has benefits in that Expr-using code can store Exprs by value instead
of via a pointer.

The cost is some additional complexity in functions that deal with Exprs at
a low level.  All current users (calc_bc_dist, x86) have been updated.

Instead of operator-based constructors (which are prone to template overload
issues and are limited to just two terms), a template-based ExprBuilder is
provided that allows construction of Expr using a natural ADD(a, b, c) syntax.
Up to 6 terms are currently allowed.

The terms are passed to ExprBuilder by const reference.  This works well for
all except for auto_ptrs, which can't be passed as const references.  We work
around this by using a single term Expr constructor in those cases, but that
isn't ideal.

The test_expr very basic test has been removed and replaced with a fairly
comprehensive set of unit tests.

In Value, m_abs is now a scoped_ptr instead of a raw pointer.
Value::add_abs() has been rewritten to take a const Expr&.