Diagnostic: Keep a SourceManager reference.
While this forces only a single SourceManager be used for all reporting
through a single Diagnostic class, this is looking to not be an issue in
yasm.  Keeping the reference in Diagnostic will reduce the need to keep
references to SourceManager (or pass it) throughout the rest of the
code.