Optimizer: fix off-by-1 error.
This could cause output assertions or infinite loops on certain inputs.

The optimizer uses an interval tree with bytecode indexes.  The bytecode
indices used as the endpoints in the tree should be everything inside of
the bounding labels of the relative expression term, but instead were
missing the first bytecode and including one bytecode after the end
label.