elf: Force relocations to be generated for PC-relative global symbols.
Formerly, if the target symbol was in the same section as the call/jmp,
a relocation would not be emitted.  While correct, this apparently
causes issues with shared library objects.

Add Object::Options structure to hold the flag that enables this
behavior so other object formats can utilize it if needed.