x86: Fix missed imm8s optimization.
Instructions such as "and esi, 0xfffffff7" take an unsigned 32-bit
value.  This value must be treated as unsigned for the correct
relocations and warnings to be generated.  However, such values can be
optimized to a signed 8-bit form (for the above example, "and esi, -9").

Detect this case and perform the optimization by sign-extending from the
immediate value size.