Clean up BytecodeContainer/Section relationship.
Instead of providing a BytecodeContainer::AsSection() virtual function
that is overridden by Section, save a Section* in BytecodeContainer and
provide a getSection() function to retrieve it.

Also move the stored Object* from BytecodeContainer to Section.

This fixes the usage of AsSection() in object formats which previously
had issues with nested BytecodeContainers (e.g.  MultipleBytecode).