Import googlemock 1.6.0.
Both googlemock and googletest are no longer distributed in a way
that they can be installed.

This necessitated a revamp of compile flags handling to only apply
yasm compile flags to yasm executables/libraries and not gmock ones.
The magic to do this is in the YASM_SET_COMPILE_FLAGS() macro in
cmake/modules/YasmMacros.cmake.