nasm: Optimize times X d* and res*.
Detect if we have a single value (for db, etc) and are in a TIMES and
do AppendFill instead of AppendData and MultipleBytecode.

Similarly for resb and kin, look for the TIMES case and multiply up to
generate a single AppendSkip instead of MultipleBytecode.  Also change
regular resb to use AppendSkip.

This results in significantly fewer bytecodes generated for standard
macros such as STRUC, ISTRUC, AT, IEND, ALIGN, and BALIGN.